+++ lizard_auth_server/admin.py
@@ -83,21 +83,15 @@ class UserProfileAdmin(admin.ModelAdmin):
             'fields': ['user',
                        'first_name',
                        'last_name',
-                       'portals',
+                   ]}),
+        (ugettext_lazy('V1 authorization'), {
+            'fields': ['portals',
                        'organisations',
                        'roles',
                    ]}),
         (ugettext_lazy('Dates'), {
             'fields': ['created_at',
                        'updated_at',
-                   ]}),
-        (ugettext_lazy('Personal data'), {
-            'fields': ['title',
-                       'street',
-                       'postal_code',
-                       'town',
-                       'phone_number',
-                       'mobile_phone_number',
                    ]}),
     )