+++ lizard_auth_server/apps.py
@@ -7,4 +7,4 @@ class MyAppConfig(AppConfig):
 
     def ready(self):
         # Enable the signals
-        from lizard_auth_server.signal_handlers import create_user_profile  # noqa
+        from lizard_auth_server.signal_handlers import create_user_profile  # NOQA