+++ lizard_auth_server/conf.py
@@ -21,3 +21,4 @@ class LizardAuthServerAppConf(AppConf):
     JWT_ALGORITHM = 'HS256'
     JWT_EXPIRATION_DELTA = timedelta(seconds=300)
     DIRTY_HARDCODED_PASSWORD = 'dirtyhardcodedpassword'
+    ACCOUNT_ACTIVATION_DAYS = 45  # A month of vacation + some margin