+++ lizard_auth_server/urls.py
@@ -41,6 +41,8 @@ urlpatterns = patterns(
     url(r'^admin/', include(admin.site.urls)),
     url(r'^i18n/', include('django.conf.urls.i18n')),
 
+    # Version 1 API
+    #
     # /api/ and /sso/api/ URLs are mainly used for internal
     # communication between the servers Note: these are referred to by
     # lizard-auth-client: change them in both places
@@ -68,6 +70,8 @@ urlpatterns = patterns(
         views_api.UserOrganisationRolesView.as_view(),
         name='lizard_auth_server.api.user_organisation_roles'),
 
+    # Version 1 views
+    #
     # SSO URLs for use by visitors Note: these are referred to by
     # lizard-auth-client: change them in both places
     url(r'^sso/portal_action/$',
@@ -79,7 +83,6 @@ urlpatterns = patterns(
     url(r'^sso/authorize/$',
         views_sso.AuthorizeView.as_view(),
         name='lizard_auth_server.sso.authorize'),
-
     # SSO URLs for use by other webservers Note: these are referred to
     # by lizard-auth-client: change them in both places
     url(r'^sso/api/request_token/$',
@@ -89,28 +92,40 @@ urlpatterns = patterns(
         views_sso.VerifyView.as_view(),
         name='lizard_auth_server.sso.api.verify'),
 
-    # V2 API
+    # Version 2 API
+    #
+    # API calls
     url(r'^api2/$',
         views_api_v2.StartView.as_view(),
         name='lizard_auth_server.api_v2.start'),
-    url(r'^api2/login/$',
-        views_api_v2.LoginView.as_view(),
-        name='lizard_auth_server.api_v2.login'),
     url(r'^api2/check_credentials/$',
         views_api_v2.CheckCredentialsView.as_view(),
         name='lizard_auth_server.api_v2.check_credentials'),
+    url(r'^api2/organisations/$',
+        views_api_v2.OrganisationsView.as_view(),
+        name='lizard_auth_server.api_v2.organisations'),
+    url(r'^api2/new_user/$',
+        views_api_v2.NewUserView.as_view(),
+        name='lizard_auth_server.api_v2.new_user'),
+    # Views for visitors
+    url(r'^api2/login/$',
+        views_api_v2.LoginView.as_view(),
+        name='lizard_auth_server.api_v2.login'),
     url(r'^api2/logout/$',
         views_api_v2.LogoutView.as_view(),
         name='lizard_auth_server.api_v2.logout'),
     url(r'^api2/logout_redirect_back_to_portal/$',
         views_api_v2.LogoutRedirectBackView.as_view(),
         name='lizard_auth_server.api_v2.logout_redirect_back'),
-    url(r'^api2/new_user/$',
-        views_api_v2.NewUserView.as_view(),
-        name='lizard_auth_server.api_v2.new_user'),
-    url(r'^api2/organisations/$',
-        views_api_v2.OrganisationsView.as_view(),
-        name='lizard_auth_server.api_v2.organisations'),
+    url(r'^api2/activate/' +
+        '(?P<user_id>[^/]+)/' +
+        '(?P<sso_key>[^/]+)/' +
+        '(?P<message>[^/]+)$',
+        views_api_v2.ActivateAndSetPasswordView.as_view(),
+        name='lizard_auth_server.api_v2.activate-and-set-password'),
+    url(r'^api2/activated/(?P<portal_pk>[^/]+)/$',
+        views_api_v2.ActivatedGoToPortalView.as_view(),
+        name='lizard_auth_server.api_v2.activated-go-to-portal'),
 
     # Override django-auth's default login/logout URLs
     # Note: ensure LOGIN_URL isn't defined in the settings
@@ -187,7 +202,7 @@ urlpatterns = patterns(
         },
         name='password_reset_complete'
     ),
-    # URLs for user invitation / activation
+    # v1 URLs for user invitation / activation
     url(
         r'^invite/$',
         views.InviteUserView.as_view(),