+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog of lizard-auth-server
 2.9 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Removed virtually unused address/phone/title fields on user profile. The new
+  v2 api won't use them anyway.
+
+- Users created through the ``/api2/new_user/`` API call now get an email with
+  an activation link.
 
 
 2.8 (2016-10-04)